- updated shad Dialog - removed extra docs dir - refactored settings page and tabs - added null check fix for unit tests to SideBar trigger icon - refactored the password validation logic into a util file - removed unnecessary awaits from tests - removed the svelte llm instructions in favor of the medium svelte5 doc in `docs`
