+++ package.json
@@ -21,6 +21,7 @@
   "devDependencies": {
     "@eslint/compat": "^1.2.5",
     "@eslint/js": "^9.18.0",
+    "@lucide/svelte": "^0.482.0",
     "@playwright/test": "^1.50.1",
     "@sveltejs/adapter-auto": "^4.0.0",
     "@sveltejs/kit": "^2.16.0",
@@ -29,7 +30,7 @@
     "@testing-library/svelte": "^5.2.7",
     "@types/node": "^22.13.5",
     "autoprefixer": "^10.4.20",
-    "bits-ui": "^1.3.2",
+    "bits-ui": "^1.3.10",
     "clsx": "^2.1.1",
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
