+++ src/lib/components/ui/dialog/dialog-description.svelte
@@ -2,13 +2,15 @@
   import { Dialog as DialogPrimitive } from 'bits-ui';
   import { cn } from '$lib/utils.js';
 
-  let { ref = $bindable(null), class: className, children, ...restProps } = $props();
+  let {
+    ref = $bindable(null),
+    class: className,
+    ...restProps
+  }: DialogPrimitive.DescriptionProps = $props();
 </script>
 
 <DialogPrimitive.Description
   bind:ref
   class={cn('text-sm text-muted-foreground', className)}
   {...restProps}
->
-  {@render children?.()}
-</DialogPrimitive.Description>
+/>
