+++ src/lib/components/ui/dialog/dialog-header.svelte
@@ -1,7 +1,14 @@
 <script lang="ts">
+  import type { HTMLAttributes } from 'svelte/elements';
+  import type { WithElementRef } from 'bits-ui';
   import { cn } from '$lib/utils.js';
 
-  let { ref = $bindable(null), class: className, children, ...restProps } = $props();
+  let {
+    ref = $bindable(null),
+    class: className,
+    children,
+    ...restProps
+  }: WithElementRef<HTMLAttributes<HTMLDivElement>> = $props();
 </script>
 
 <div
