+++ src/lib/components/ui/dialog/dialog-title.svelte
@@ -2,13 +2,15 @@
   import { Dialog as DialogPrimitive } from 'bits-ui';
   import { cn } from '$lib/utils.js';
 
-  let { ref = $bindable(null), class: className, children, ...restProps } = $props();
+  let {
+    ref = $bindable(null),
+    class: className,
+    ...restProps
+  }: DialogPrimitive.TitleProps = $props();
 </script>
 
 <DialogPrimitive.Title
   bind:ref
   class={cn('text-lg font-semibold leading-none tracking-tight', className)}
   {...restProps}
->
-  {@render children?.()}
-</DialogPrimitive.Title>
+/>
