+++ src/lib/components/ui/dialog/index.ts
@@ -1,6 +1,37 @@
-export { default as Content } from './dialog-content.svelte';
-export { default as Description } from './dialog-description.svelte';
-export { default as Footer } from './dialog-footer.svelte';
-export { default as Header } from './dialog-header.svelte';
-export { default as Title } from './dialog-title.svelte';
-export { default as Root } from './dialog.svelte';
+import { Dialog as DialogPrimitive } from 'bits-ui';
+
+import Content from './dialog-content.svelte';
+import Description from './dialog-description.svelte';
+import Footer from './dialog-footer.svelte';
+import Header from './dialog-header.svelte';
+import Overlay from './dialog-overlay.svelte';
+import Title from './dialog-title.svelte';
+
+const Root = DialogPrimitive.Root;
+const Trigger = DialogPrimitive.Trigger;
+const Close = DialogPrimitive.Close;
+const Portal = DialogPrimitive.Portal;
+
+export {
+  Close,
+  Content,
+  Description,
+  //
+  Root as Dialog,
+  Close as DialogClose,
+  Content as DialogContent,
+  Description as DialogDescription,
+  Footer as DialogFooter,
+  Header as DialogHeader,
+  Overlay as DialogOverlay,
+  Portal as DialogPortal,
+  Title as DialogTitle,
+  Trigger as DialogTrigger,
+  Footer,
+  Header,
+  Overlay,
+  Portal,
+  Root,
+  Title,
+  Trigger,
+};
